Move AMP from contrib to core
