+++ src/operator/all_finite.cc
